v0.0.16: Add `Viewer` for image show
* Add `Viewer` for image-show

* Add .`with_ixx()` to simplify model building with `Options`

* Update `MinOptMax`

+++ src/models/blip.rs
@@ -56,8 +56,8 @@ impl Blip {
         let xs_ = X::apply(&[
             Ops::Resize(
                 xs,
-                self.height.opt as u32,
-                self.width.opt as u32,
+                self.height.opt() as u32,
+                self.width.opt() as u32,
                 "Bilinear",
             ),
             Ops::Normalize(0., 255.),
@@ -146,10 +146,10 @@ impl Blip {
     }
 
     pub fn batch_visual(&self) -> usize {
-        self.batch_visual.opt as usize
+        self.batch_visual.opt()
     }
 
     pub fn batch_textual(&self) -> usize {
-        self.batch_textual.opt as usize
+        self.batch_textual.opt()
     }
 }
